Giving up for now on alignment dreams
aligning the output pads to the right while the input nodes are aligned to the right is
either not possible, or too complicated for my tiny brain. I'll try asking for help on
this later, but I can't waste too much more time on this. It looks terrible, but oh well.
It will have to do for now.