Fix jittering on scroll
I'm being kind of dumb and sometimes creating a local variable in function calls so I
don't have to have the full dereference each time, and I forgot to save back the origin
field when it gets updated. The whole thing could use some rethinking, I was just trying
to go for the sort of calling convention DearImGUI uses, might have to look at how they
deal with this sort of thing.